refactor: 예약시간 DB 접근 로직 Repository으로 이동
